py-opcua-collector/requirements.txt
aiofiles==22.1.0
aiosqlite==0.17.0
asyncua==1.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
idna==3.4
influxdb-client==1.34.0
msgpack==1.0.4
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
reactivex==4.0.4
requests==2.28.1
six==1.16.0
sortedcontainers==2.4.0
typing-extensions==4.4.0
urllib3==1.26.12
